[CHORE](foundation-api): Split ask handler by concern
Replace the 570-line `routes/ask.rs` with a `routes/ask/` module
where each submodule owns its own inline tests:

    routes/ask/
        mod.rs           -- module declarations + pub(crate) use
        error.rs         -- AskProxyError + ChromaError impl + its test
        merge.rs         -- merge_user helper + four unit tests
        handler.rs       -- the axum `ask` handler + seven integration
                            tests that exercise it end-to-end
        test_support.rs  -- shared #[cfg(test)] fixtures: TestAuth,
                            stub-Modal server, FoundationApiServer
                            builder, post_ask helper, body reader

Folds the two test-only auth structs (FakeAuth, UnauthorizedAuth)
into a single `TestAuth` enum to drop boilerplate, and reuses the
stub-server spawn helper across both the canned-success and the
custom-status tests in handler.rs.

Production behavior unchanged.
